+++ app/Admin/Actions/Paper/DeleteJudgement.php
@@ -3,11 +3,13 @@
 namespace App\Admin\Actions\Paper;
 
 use App\Judgement;
+use App\StudentScore;
 use Encore\Admin\Actions\RowAction;
 use Illuminate\Database\Eloquent\Model;
 
 class DeleteJudgement extends RowAction
 {
+    // 判断题-删除
     public $name = '删除';
 
     public function handle(Model $model)
@@ -20,7 +22,13 @@ class DeleteJudgement extends RowAction
 
         if ($model instanceof Judgement) {
             try {
-                $model->delete();
+                // 该试卷已录入学生答案则不可删除
+                $stu_score = StudentScore::where('paper_id', $model->paper_id)->value('id');
+                if (!$stu_score) {
+                    $model->delete();
+                } else {
+                    return $this->response()->error("{$trans['failed']} : <br>该试卷已录入学生答案，不可删除题目！<br>如非要删除，请先删除该试卷下所有学生答案！");
+                }
             } catch (\Exception $exception) {
                 return $this->response()->error("{$trans['failed']} : {$exception->getMessage()}");
             }
